Added authorization (edit and delete)

+++ backend/routes/posts.js
@@ -55,6 +55,7 @@ router.post(
 );
 
 // edit existing post
+// protected from editing by unauthorized users
 router.put(
   "/:id",
   checkAuth,
@@ -71,8 +72,13 @@ router.put(
       content: req.body.content,
       imagePath: imagePath
     });
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-      res.status(200).json({message: "Update successful!"});
+    // updateOne takes arguments of post id and creator id
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if(result.nModified > 0) {
+        res.status(200).json({message: "Update successful!"});
+      } else {
+        res.status(401).json({message: "User not authorized to edit"});
+      }
     });
   }
 );
@@ -115,9 +121,12 @@ router.get("/:id", (req, res, next) => {
 
 // delete post
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(req.params.id);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if(result.n > 0) {
+      res.status(200).json({message: "Deletion successful!"});
+    } else {
+      res.status(401).json({message: "Not authorized!"});
+    }
   });
 });
